ICU4N.Impl.UCharacterProperty: Eliminated most casting by changing "which" parameters from int to UProperty

+++ src/ICU4N/Globalization/UCharacter.cs
@@ -5702,7 +5702,7 @@ namespace ICU4N
         /// <stable>ICU 2.6</stable>
         public static bool HasBinaryProperty(int ch, UProperty property)
         {
-            return UCharacterProperty.Instance.HasBinaryProperty(ch, (int)property);
+            return UCharacterProperty.Instance.HasBinaryProperty(ch, property);
         }
 
         /// <icu/>
